+++ tests/testthat/test-pipe.R
@@ -1,0 +1,7 @@
+test_that("multiple expectations in a pipe chain work", {
+  mtcars %>%
+    with_testdata(expect_base(mpg, TRUE)) %>%
+    mutate(mpg = NA) %>%
+    with_testdata(expect_base(mpg, FALSE))
+})
+